+++ benchmarks/Benchmark.hs
@@ -8,6 +8,7 @@ import qualified System.Random.PCG as PCG
 import qualified System.Random.PCG.Fast as F
 import qualified System.Random.PCG.Single as S
 import qualified System.Random.PCG.Unique as U
+import           System.Random.PCG.Class (fastUniformB)
 import qualified System.Random.Mersenne as M
 
 benchIO :: String -> IO a -> Benchmark
@@ -28,12 +29,14 @@ main = do
     --   ]
     [ bgroup "pcg-fast"
       [ benchIO "Word32" (F.uniform pcgF :: IO Word32)
-      , benchIO "Word32R" (F.uniformR (0,10000) pcgF :: IO Word32)
-      , benchIO "Word32B" (F.uniformB 10000 pcgF :: IO Word32)
+      -- , benchIO "Word32R" (F.uniformR (0,10000) pcgF :: IO Word32)
+      -- , benchIO "Word32B" (F.uniformB 10000 pcgF :: IO Word32)
       , benchIO "Word64B" (F.uniformB 10000 pcgF :: IO Word64)
       , benchIO "IntBViaFloat" (viaFloat 10000 pcgF :: IO Int)
-      , benchIO "rs-Word64" (F.uniform pcgF :: IO Word64)
-      , benchIO "rs-Double" (F.uniform pcgF :: IO Double)
+      , benchIO "fast" (fastUniformB 10000 pcgF :: IO Int)
+      , benchIO "fastW32" (fastUniformB 1000 pcgF :: IO Word32)
+      -- , benchIO "rs-Word64" (F.uniform pcgF :: IO Word64)
+      -- , benchIO "rs-Double" (F.uniform pcgF :: IO Double)
       ]
     -- , bgroup "pcg-single"
     --   [ benchIO "Word32" (S.uniform pcgS :: IO Word32)
@@ -41,12 +44,12 @@ main = do
     -- , bgroup "pcg-unique"
     --   [ benchIO "Word32" (U.uniform pcgU :: IO Word32)
     --   ]
-    , bgroup "mwc"
-      [ benchIO "Word64" (MWC.uniform mwc :: IO Word64)
-      , benchIO "Word32R" (MWC.uniformR (0,10000) mwc :: IO Word32)
-      , benchIO "Double" (MWC.uniform mwc :: IO Double)
+    -- , bgroup "mwc"
+      -- [ benchIO "Word64" (MWC.uniform mwc :: IO Word64)
+      -- , benchIO "Word32R" (MWC.uniformR (0,10000) mwc :: IO Word32)
+      -- , benchIO "Double" (MWC.uniform mwc :: IO Double)
       -- , bench "Int"    (uniform mwc :: IO Int)
-      ]
+      -- ]
     -- , bgroup "random"
     --   [ benchIO "Word32" (R.randomIO >>= evaluate :: IO Word32)
     --   , benchIO "Double" (R.randomIO >>= evaluate :: IO Double)