+++ src/System/Random/PCG/Fast.hs
@@ -39,6 +39,7 @@ module System.Random.PCG.Fast
   ( -- * Gen
     Gen, GenIO, GenST
   , create, createSystemRandom, initialize, withSystemRandom
+  , fastUniformB
 
     -- * Getting random numbers
   , Variate (..)