+++ src/System/Random/PCG/Single.hs
@@ -37,6 +37,7 @@ module System.Random.PCG.Single
   ( -- * Gen
     Gen, GenIO, GenST
   , create, createSystemRandom, initialize, withSystemRandom
+  , fastUniformB
 
     -- * Getting random numbers
   , Variate (..)