+++ src/System/Random/PCG/Unique.hs
@@ -36,6 +36,7 @@ module System.Random.PCG.Unique
   ( -- * Gen
     Gen
   , create, createSystemRandom, initialize, withSystemRandom
+  , fastUniformB
 
     -- * Getting random numbers
   , Variate (..)