+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+### 0.1.1.0
+
+* Add uniformB, a function to generate a bounded random number in [0,b)
+  range. This preforms significantly faster than uniformR (0,b-1).
+* Add type specific versions for uniformR and uniformB.
+
 ### 0.1.0.1
 
 * Fix bug when dealing with `Word` and `Int` types.